Fix ajax include + add automating itsm module

+++ module/admin_itsm/function_itsm.php
@@ -25,17 +25,19 @@
  * 
  * @return boolean
  */
-function upload_file($file, $dir="uploaded_file/"){
+function upload_file($file, $dir="uploaded_file"){
     $old_file = get_itsm_var("itsm_file");
-    unlink($old_file);
-
+    if(file_exists($old_file)){
+        unlink($old_file);
+    }
     if(isset($file)){
-        $target_file = __DIR__."/".$dir.basename($file["name"]);
+        $target_file = __DIR__."/".$dir."/".basename($file["name"]);
         if(move_uploaded_file($file["tmp_name"], $target_file)){
             return true;
         }
         return false;
     }
+    
     return false;
 }
 
@@ -132,6 +134,131 @@ function report_itsm($detail, $descr){
     }else return false;
 }
 
+/**
+ * 
+ * This function format the data to become able to use them the same 
+ * way as monitoring_ged. 
+ * 
+ */
+function get_all_events(){
+    global $database_ged;
+    $events = array();
+    $sql = "SELECT id FROM nagios_queue_active";
+    $result = sqlrequest($database_ged, $sql);
+    while ($row = mysqli_fetch_assoc($result)){
+        array_push($events,$row["id"].":nagios");
+    }
+    $result = null;
+    $sql = "SELECT id FROM snmptrap_queue_active";
+    $result = sqlrequest($database_ged, $sql);
+    while ($row = mysqli_fetch_assoc($result)){
+        array_push($events,$row["id"].":snmptrap");
+    }
 
+    return $events;
+}
 
+/**
+ * Function from ged_functions.php of monitoring_ged module
+ * Thoses function is use when the automation his enable.
+ *  
+ */
+function acknowledge($selected_events, $queue)
+{
+	global $array_ged_queues;
+	global $database_ged;
+	global $array_ged_packets;
+	global $path_ged_bin;
+	global $array_serv_system;
+
+	if(!in_array($queue,$array_ged_queues)) { $queue=$array_ged_queues[0]; }
+	
+	if(exec($array_serv_system["Ged agent"]["status"])==NULL) {
+		return message(0," : ged daemon must be dead","critical");
+	}
+
+	$owner = $_COOKIE['user_name']."@".getenv("SERVER_NAME");
+
+	foreach ($selected_events as $value) {
+		$value_parts = explode(":", $value);
+		$id = $value_parts[0];
+		$ged_type = $value_parts[1];
+		if($ged_type == "nagios"){ $ged_type_nbr = 1; }
+		if($ged_type == "snmptrap"){ $ged_type_nbr = 2; }
+
+		$event_to_delete = [];
+		array_push($event_to_delete, $value);
+
+		$sql = "SELECT * FROM ".$ged_type."_queue_".$queue." WHERE id = ?";
+		$result = sqlrequest($database_ged, $sql, false, array("s",(string)$id));
+		$event = mysqli_fetch_assoc($result);
+
+		$ged_command = "-update -type $ged_type_nbr ";
+		foreach ($array_ged_packets as $key => $value) {
+			if($value["type"] == true){
+				if($key == "owner"){
+					$event[$key] = $owner;
+				}
+				$ged_command .= escapeshellarg($event[$key])." ";
+			}
+		}
+		$ged_command = trim($ged_command, " ");
+		
+		shell_exec($path_ged_bin." ".$ged_command);
+		logging("ged_update",$ged_command);
+		delete($event_to_delete, $queue);
+	}
+}
+
+function delete($selected_events, $queue)
+{
+	global $array_ged_queues;
+	global $database_ged;
+	global $array_ged_packets;
+	global $path_ged_bin;
+	global $array_serv_system;
+
+	if(!in_array($queue,$array_ged_queues)) { $queue=$array_ged_queues[0]; }
+	
+	if(exec($array_serv_system["Ged agent"]["status"])==NULL) {
+		return message(0," : ged daemon must be dead","critical");
+	}
+
+	$id_list = "";
+	foreach ($selected_events as $value) {
+		$value_parts = explode(":", $value);
+		$id = $value_parts[0];
+		$ged_type = $value_parts[1];
+		$ged_type_nbr = 0;
+		if($ged_type == "nagios"){ $ged_type_nbr = 1; }
+		if($ged_type == "snmptrap"){ $ged_type_nbr = 2; }
+
+		$sql = "SELECT * FROM ".$ged_type."_queue_".$queue." WHERE id = $id";
+		$result = sqlrequest($database_ged, $sql);
+		$event = mysqli_fetch_assoc($result);
+
+		if($queue == "active"){
+			$ged_command = "-drop -type $ged_type_nbr -queue $queue ";
+			foreach ($array_ged_packets as $key => $value) {
+				if($value["key"] == true){
+					$ged_command .= escapeshellarg($event[$key])." ";
+				}
+			}
+			$ged_command = trim($ged_command, " ");
+					
+			shell_exec($path_ged_bin." ".$ged_command);
+			logging("ged_update",$ged_command);
+		} else {
+			$id_list .= $id.",";
+		}
+	}
+
+	if($queue == "history"){
+		$id_list = trim($id_list, ",");
+		$ged_command = "-drop -id ".escapeshellarg($id_list)." -queue history";
+		
+		shell_exec($path_ged_bin." ".$ged_command);
+		logging("ged_update",$ged_command);
+	}
+}
 ?>
